routes and controllers of findAll/Create people
